break up contents of rpc mod into sub-modules
