+++ crates/server/src/rpc_server/mod.rs
@@ -1,0 +1,5 @@
+mod server;
+mod types;
+
+pub use server::{ContenderRpcServer, ContenderServer};
+pub use types::*;
