+++ crates/server/src/rpc_server/types.rs
@@ -1,0 +1,146 @@
+use crate::{error::ContenderRpcError, sessions::NewSessionParams};
+use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
+use contender_cli::default_scenarios::{BuiltinOptions, BuiltinScenarioCli};
+use contender_core::{
+    alloy::{
+        network::AnyNetwork,
+        providers::{DynProvider, ProviderBuilder},
+    },
+    generator::RandSeed,
+    test_scenario::Url,
+};
+use contender_testfile::TestConfig;
+use serde::{Deserialize, Serialize};
+use std::str::FromStr;
+use tracing::debug;
+
+/// RPC parameters for adding a new contender session.
+#[derive(Clone, Debug, Deserialize)]
+pub struct AddSessionParams {
+    pub name: String,
+    pub rpc_url: Url,
+    pub test_config: Option<TestConfigSource>,
+}
+
+impl AddSessionParams {
+    pub async fn to_new_session_params(
+        self,
+        seed: RandSeed,
+    ) -> Result<NewSessionParams, ContenderRpcError> {
+        let test_config = if let Some(config) = self.test_config {
+            let provider = DynProvider::new(
+                ProviderBuilder::new()
+                    .network::<AnyNetwork>()
+                    .connect_http(self.rpc_url.clone()),
+            );
+            config
+                .to_testconfig(
+                    Some(BuiltinOptions {
+                        accounts_per_agent: None,
+                        seed,
+                        spam_rate: None,
+                    }),
+                    &provider,
+                )
+                .await?
+        } else {
+            TestConfig::from_str(include_str!("../../../../scenarios/uniV2.toml"))
+                .expect("default config should be valid")
+        };
+
+        Ok(NewSessionParams {
+            name: self.name.clone(),
+            rpc_url: self.rpc_url.clone(),
+            test_config,
+        })
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub enum TestConfigSource {
+    TomlBase64(String),
+    Json(TestConfig),
+    Builtin(BuiltinScenarioCli),
+}
+
+impl TestConfigSource {
+    pub async fn to_testconfig(
+        self,
+        builtin_options: Option<BuiltinOptions>,
+        provider: &DynProvider<AnyNetwork>,
+    ) -> Result<TestConfig, ContenderRpcError> {
+        match self {
+            TestConfigSource::TomlBase64(b64) => {
+                let bytes = BASE64.decode(b64)?;
+                debug!(
+                    "Decoded test config from base64, length {} bytes",
+                    bytes.len()
+                );
+                let config_str =
+                    String::from_utf8(bytes).map_err(ContenderRpcError::InvalidUtf8)?;
+                TestConfig::from_str(&config_str).map_err(ContenderRpcError::InvalidTestConfig)
+            }
+
+            TestConfigSource::Json(config) => Ok(config),
+
+            TestConfigSource::Builtin(builtin) => {
+                let scenario = builtin
+                    .to_builtin_scenario(provider, builtin_options.unwrap_or_default())
+                    .await
+                    .unwrap()
+                    .into();
+                Ok(scenario)
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use base64::engine::general_purpose::STANDARD as BASE64;
+    use contender_cli::default_scenarios::transfers::TransferStressCliArgs;
+    use contender_core::alloy::{
+        consensus::constants::ETH_TO_WEI,
+        primitives::{Address, U256},
+    };
+
+    #[test]
+    fn test_toml_base64_variant() {
+        let toml_content = include_str!("../../../../scenarios/uniV2.toml");
+        let b64 = BASE64.encode(toml_content);
+        let json = serde_json::json!({ "TomlBase64": b64 });
+        // println!(
+        //     "TomlBase64:\n{}\n",
+        //     serde_json::to_string_pretty(&json).unwrap()
+        // );
+
+        let source: TestConfigSource = serde_json::from_value(json).unwrap();
+        assert!(matches!(source, TestConfigSource::TomlBase64(_)));
+    }
+
+    #[test]
+    fn test_json_variant() {
+        let config =
+            TestConfig::from_str(include_str!("../../../../scenarios/uniV2.toml")).unwrap();
+        let json = serde_json::json!({ "Json": config });
+        // println!("Json:\n{}\n", serde_json::to_string_pretty(&json).unwrap());
+
+        let source: TestConfigSource = serde_json::from_value(json).unwrap();
+        assert!(matches!(source, TestConfigSource::Json(_)));
+    }
+
+    #[tokio::test]
+    async fn test_builtin_variant() {
+        let builtin =
+            TestConfigSource::Builtin(BuiltinScenarioCli::Transfers(TransferStressCliArgs {
+                amount: U256::from(ETH_TO_WEI),
+                recipient: Some(Address::ZERO),
+            }));
+        let json = serde_json::json!(builtin);
+        // println!("{}", serde_json::to_string_pretty(&json).unwrap());
+
+        let source: TestConfigSource = serde_json::from_value(json).unwrap();
+        assert!(matches!(source, TestConfigSource::Builtin(_)));
+    }
+}
